requirements.txt
tensorflow==2.15.0
# tensorflow-estimator==2.15.0
# tensorflow-io-gcs-filesystem==0.31.0
numpy==1.26.4
mediapipe==0.10.11
opencv-python==4.9.0.80
# scikit-learn==1.4.1.post1
# kaggle==1.6.14
google-cloud-storage==2.16.0
starlette==0.37.2
uvicorn==0.30.1
pydantic==2.7.1
fastapi==0.111.0